Move internal headers from include/private to src/include
