+++ src/include/hash_string.h
