+++ src/include/pmix_globals.h
@@ -22,7 +22,7 @@
 
 #include <private/autogen/config.h>
 #include <pmix/rename.h>
-#include <private/types.h>
+#include <src/include/types.h>
 
 #include <unistd.h>
 #ifdef HAVE_SYS_TYPES_H
